Introduce src/settings.rs, for dynamic config handling
