+++ src/workspace.rs
@@ -1,5 +1,6 @@
 use crate::context::*;
 use crate::language_features::rust_analyzer;
+use crate::settings::explode_string_table;
 use crate::types::*;
 use crate::util::*;
 use jsonrpc_core::Params;
@@ -10,73 +11,6 @@ use serde::Deserialize;
 use serde_json::{self, Value};
 use std::fs;
 use std::io;
-
-fn insert_value<'a, 'b, P>(
-    target: &'b mut serde_json::map::Map<String, Value>,
-    mut path: P,
-    local_key: String,
-    value: Value,
-) -> Result<(), String>
-where
-    P: Iterator<Item = &'a str>,
-    P: 'a,
-{
-    match path.next() {
-        Some(key) => {
-            let maybe_new_target = target
-                .entry(key)
-                .or_insert_with(|| Value::Object(serde_json::Map::new()))
-                .as_object_mut();
-
-            if maybe_new_target.is_none() {
-                return Err(format!(
-                    "Expected path {:?} to be object, found {:?}",
-                    key, &maybe_new_target,
-                ));
-            }
-
-            insert_value(maybe_new_target.unwrap(), path, local_key, value)
-        }
-        None => match target.insert(local_key, value) {
-            Some(old_value) => Err(format!("Replaced old value: {:?}", old_value)),
-            None => Ok(()),
-        },
-    }
-}
-
-// Take flattened tables like "a.b = 1" and produce "{"a":{"b":1}}".
-pub fn explode_string_table(raw_settings: &toml::value::Table) -> Value {
-    let mut settings = serde_json::Map::new();
-
-    for (raw_key, raw_value) in raw_settings.iter() {
-        let mut key_parts = raw_key.split('.');
-        let local_key = match key_parts.next_back() {
-            Some(name) => name,
-            None => {
-                warn!("Got a setting with an empty local name: {:?}", raw_key);
-                continue;
-            }
-        };
-
-        let value: Value = match raw_value.clone().try_into() {
-            Ok(value) => value,
-            Err(e) => {
-                warn!("Could not convert setting {:?} to JSON: {}", raw_value, e,);
-                continue;
-            }
-        };
-
-        match insert_value(&mut settings, key_parts, local_key.into(), value) {
-            Ok(_) => (),
-            Err(e) => {
-                warn!("Could not set {:?} to {:?}: {}", raw_key, raw_value, e);
-                continue;
-            }
-        }
-    }
-
-    Value::Object(settings)
-}
 
 pub fn did_change_configuration(params: EditorParams, ctx: &mut Context) {
     let default_settings = toml::value::Table::new();
